feat(web): 승인 창구를 `/approvals` 한 곳으로 통합
승인이 필요한 것이 세 종류인데 화면이 전부 흩어져 있었고 셋 다 사이드바에 없었다 —
스킬 draft 는 `/skill-library` 의 서브탭, MCP draft 는 `설정 → 커넥터` 의 서브탭,
에이전트 작업 HITL 은 채팅 인라인(연결이 끊기면 유실)뿐이었다. 실제 사용자가
"승인을 어디에서 해야할 지를 모르겠어"에서 막혔다.

- 신규 `/approvals` — 에이전트 작업 HITL · 스킬 draft · MCP draft 를 한 페이지에.
  HITL 은 작업이 실제로 멈춰 대기 중이라 맨 위. Custom Agent 는 승인 개념이 없다는
  사실을 명시해(설치 즉시 활성) "왜 여기 없지"를 없앤다.
- 사이드바에 항목 추가 + 배지를 **세 종류 합산**으로 변경(기존엔 HITL 만 세어
  `/agent-tasks` 에 붙어 있었다). 축 하나가 실패해도 나머지 합계는 보이도록 개별 fallback.
- 구 진입점은 링크로 대체 — `/skill-library` Draft 탭 제거, 커넥터 Draft 탭 제거,
  Git import 성공 시 `/approvals` 로 이동.

⚠️ 부수 수정 — 커넥터 Draft 탭은 `status === "pending"` 일 때만 승인 버튼을 렌더했는데
`GET /api/mcp/servers/drafts` 는 `WHERE status='draft'` 로만 조회한다. 즉 **그 화면에서는
승인 버튼이 한 번도 보인 적이 없다**(조건이 영구 거짓). 탭 자체를 걷어내며 해소됐고,
새 `McpDrafts` 는 상태 필터 없이 목록 전부에 버튼을 단다.

위험 표시는 정적 룰의 error 만 근거로 한다 — LLM audit findings 는 warn 강등이라
차단 근거가 아니다(2026-08-24 오탐 수정과 정합).
